+++ tests/converge.py
@@ -1,0 +1,3 @@
+from server_setup import postgres_db
+
+postgres_db(database="t_db", owner="t_owner")